Library File Transpiler
- Remove unnecessary interfaces and abstract classes
- Add documentation
